requirements.txt
asgiref==3.2.10
Django==3.0.8
django-crontab==0.7.1
djangorestframework==3.11.0
djangorestframework-simplejwt==4.4.0
psycopg2-binary==2.8.5
PyJWT==1.7.1
python-dotenv==0.14.0
pytz==2020.1
sqlparse==0.3.1
